services/analytics-service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
gunicorn==21.2.0
python-decouple==3.8
requests==2.31.0
python-dateutil==2.8.2
django-extensions==3.2.3
whitenoise==6.6.0
dj-database-url==2.1.0
django-filter==23.5
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
